third_party/WebKit/LayoutTests/fast/dom/HTMLTemplateElement/cycles-in-shadow-expected.txt

Test that cycle detection traverses over both templates and shadow roots

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS shadowDiv.appendChild(outerDiv) threw exception HierarchyRequestError: A Node was inserted somewhere it doesn't belong..
PASS successfullyParsed is true

TEST COMPLETE
